MACOSX: Put documentation links in Help menu
The local documents are not currently internationalised simply
because the internationalised resources are not put into the right
places (NSBundle will handle this automatically when they are);
Trac#10464 is a tracking bug for this outstanding issue.

Also, in order to reduce the complexity of showing the licensing
information, the licensing files are now collapsed into a single
COPYING file, and non-license information (like the instructions
from GNU on how to apply the GPL to a software project) has been
stripped. This is of benefit to everyone, since some ports were
not including all the correct licensing information since they
weren’t updated when new license files were added to the codebase.

Fixes Trac#10437.